Event Managment System/src/main/java/com/example/demo/message/MessageServiceImp.java
package com.example.demo.message;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List; 

@Service
public class MessageServiceImp implements MessageService {

    @Autowired
    private MessageRepository messageRepository; 

    @Override
    public Message saveMessage(Message message) {
        return messageRepository.save(message);
    }

    @Override
    public List<Message> getAllMessages() {
        return messageRepository.findAll();
    }

    @Override
    public Message getMessageById(int id) {
        return messageRepository.findById(id)
                .orElseThrow(() -> new RuntimeException("Message not found with ID: " + id));
    }


    @Override
    public void deleteMessage(int id) {
        messageRepository.deleteById(id);
    }
}
